remove unnecessary ReceiptAction for local receipts
