+++ Modules/Sources/Yosemite/Model/Mocks/ActionHandlers/MockReceiptActionHandler.swift
@@ -8,10 +8,6 @@ struct MockReceiptActionHandler: MockActionHandler {
     let storageManager: StorageManagerType
 
     func handle(action: ActionType) {
-        switch action {
-        case .loadReceipt(_, let onCompletion):
-            onCompletion(.failure(ReceiptStoreError.fileNotFound))
-        default: unimplementedAction(action: action)
-        }
+        unimplementedAction(action: action)
     }
 }
