+++ WooCommerce/Classes/Yosemite/AuthenticatedState.swift
@@ -130,9 +130,7 @@ class AuthenticatedState: StoresManagerState {
                                     cardReaderConfigProvider: ServiceLocator.cardReaderConfigProvider),
             ReceiptStore(dispatcher: dispatcher,
                          storageManager: storageManager,
-                         network: network,
-                         receiptPrinterService: ServiceLocator.receiptPrinterService,
-                         fileStorage: PListFileStorage()),
+                         network: network),
             AnnouncementsStore(dispatcher: dispatcher,
                                storageManager: storageManager,
                                network: network,
